test_samplecode.py
import unittest
import samplecode

class TestSamplecode(unittest.TestCase):
<<<<<<< HEAD
=======
    def test_add(self):
        # TODO: Implement test for add
        # Expected return value of type int
        result = samplecode.add({'name': 'a', 'type': 'int'}=None, {'name': 'b', 'type': 'int'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_subtract(self):
        # TODO: Implement test for subtract
        # Expected return value of type float
        result = samplecode.subtract({'name': 'a', 'type': 'float'}=None, {'name': 'b', 'type': 'float'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

>>>>>>> c15e9c9 (Added parameter and return type to json)
    def test_multiply(self):
        # TODO: Implement test for multiply
        # Expected return value of type int
        result = samplecode.multiply({'name': 'self', 'type': 'Any'}=None, {'name': 'a', 'type': 'int'}=None, {'name': 'b', 'type': 'int'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_divide(self):
        # TODO: Implement test for divide
        # Expected return value of type float
        result = samplecode.divide({'name': 'self', 'type': 'Any'}=None, {'name': 'a', 'type': 'int'}=None, {'name': 'b', 'type': 'int'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_power(self):
        # TODO: Implement test for power
        # Expected return value of type int
        result = samplecode.power({'name': 'self', 'type': 'Any'}=None, {'name': 'base', 'type': 'int'}=None, {'name': 'exponent', 'type': 'int'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_square_root(self):
        # TODO: Implement test for square_root
        # Expected return value of type float
        result = samplecode.square_root({'name': 'self', 'type': 'Any'}=None, {'name': 'number', 'type': 'float'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_Calculator_multiply(self):
        # TODO: Implement test for Calculator.multiply
        # Expected return value of type int
        obj = samplecode.Calculator()  # Create an instance of the class
        result = obj.multiply({'name': 'a', 'type': 'int'}=None, {'name': 'b', 'type': 'int'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_Calculator_divide(self):
        # TODO: Implement test for Calculator.divide
        # Expected return value of type float
        obj = samplecode.Calculator()  # Create an instance of the class
        result = obj.divide({'name': 'a', 'type': 'int'}=None, {'name': 'b', 'type': 'int'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_ScientificCalculator_power(self):
        # TODO: Implement test for ScientificCalculator.power
        # Expected return value of type int
        obj = samplecode.ScientificCalculator()  # Create an instance of the class
        result = obj.power({'name': 'base', 'type': 'int'}=None, {'name': 'exponent', 'type': 'int'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result

    def test_ScientificCalculator_square_root(self):
        # TODO: Implement test for ScientificCalculator.square_root
        # Expected return value of type float
        obj = samplecode.ScientificCalculator()  # Create an instance of the class
        result = obj.square_root({'name': 'number', 'type': 'float'}=None)
        self.assertEqual(result, None)  # Replace None with the expected result


if __name__ == "__main__":
    unittest.main()
